// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type SiteSettings = {
  _id: string;
  _type: "siteSettings";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  companyName?: string;
  phone?: string;
  email?: string;
  address?: {
    street?: string;
    city?: string;
    province?: string;
    postal?: string;
    country?: string;
  };
  socialLinks?: Array<{
    platform?: "LinkedIn" | "Twitter/X" | "Facebook" | "Instagram" | "YouTube";
    url?: string;
    _key: string;
  }>;
  defaultSeo?: {
    metaTitle?: string;
    metaDescription?: string;
    ogImage?: {
      asset?: SanityImageAssetReference;
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    };
  };
  stats?: Array<{
    value?: string;
    label?: string;
    _key: string;
  }>;
  footerBlurb?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
};